% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.convert_monthly_to_annual_estoque}
\alias{.convert_monthly_to_annual_estoque}
\title{Converter dados mensais de estoque para anuais (último valor de dez/ano)}
\usage{
.convert_monthly_to_annual_estoque(df)
}
\arguments{
\item{df}{Tibble com dados de estoque (coluna Ano como data)}
}
\value{
Tibble com dados anuais
}
\description{
Converter dados mensais de estoque para anuais (último valor de dez/ano)
}
