% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.ensure_hierarchical_cnae_columns}
\alias{.ensure_hierarchical_cnae_columns}
\title{Garantir que códigos CNAE hierárquicos estejam presentes}
\usage{
.ensure_hierarchical_cnae_columns(df, nivel_cnae)
}
\arguments{
\item{df}{Tibble com dados}

\item{nivel_cnae}{Nível CNAE}
}
\value{
Tibble verificado
}
\description{
Garantir que códigos CNAE hierárquicos estejam presentes
}
