% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.filter_cnae_columns_by_level}
\alias{.filter_cnae_columns_by_level}
\title{Filtrar colunas CNAE específicas baseado no nível CNAE selecionado}
\usage{
.filter_cnae_columns_by_level(df, nivel_cnae)
}
\arguments{
\item{df}{Tibble com dados}

\item{nivel_cnae}{Nível CNAE}
}
\value{
Tibble filtrado
}
\description{
Filtrar colunas CNAE específicas baseado no nível CNAE selecionado
}
