% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.filter_cnae_columns_for_grouped_methods}
\alias{.filter_cnae_columns_for_grouped_methods}
\title{Filtrar colunas CNAE específicas dos métodos agrupados}
\usage{
.filter_cnae_columns_for_grouped_methods(df, nivel_agregacao)
}
\arguments{
\item{df}{Tibble com dados}

\item{nivel_agregacao}{Nível de agregação}
}
\value{
Tibble filtrado
}
\description{
Filtrar colunas CNAE específicas dos métodos agrupados
}
