% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.filter_columns_by_aggregation}
\alias{.filter_columns_by_aggregation}
\title{Filtrar colunas baseado no nível de agregação geográfica}
\usage{
.filter_columns_by_aggregation(df, nivel_agregacao)
}
\arguments{
\item{df}{Tibble com dados}

\item{nivel_agregacao}{Nível de agregação}
}
\value{
Tibble filtrado
}
\description{
Filtrar colunas baseado no nível de agregação geográfica
}
