% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.validate_cnae_codes}
\alias{.validate_cnae_codes}
\title{Validar códigos CNAE e determinar nível automaticamente}
\usage{
.validate_cnae_codes(codigos_cnae)
}
\arguments{
\item{codigos_cnae}{Vetor de códigos CNAE}
}
\value{
String do nível determinado
}
\description{
Validar códigos CNAE e determinar nível automaticamente
}
