% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{.validate_cnae_level}
\alias{.validate_cnae_level}
\title{Validar e converter nível CNAE string para int da API}
\usage{
.validate_cnae_level(nivel_cnae)
}
\arguments{
\item{nivel_cnae}{String do nível CNAE ('subclasse', 'divisao', 'grupo')}
}
\value{
Valor inteiro ou NULL correspondente
}
\description{
Validar e converter nível CNAE string para int da API
}
