% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_date_bases}
\alias{get_date_bases}
\title{Obter a data da última atualização de cada base (ComexStat, CAGED, RAIS)}
\usage{
get_date_bases()
}
\value{
Tibble com uma linha por base ('Base' e 'DataUltimaAtualizacao')
}
\description{
Obter a data da última atualização de cada base (ComexStat, CAGED, RAIS)
}
