% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_estadual_agrupado}
\alias{get_estoque_emprego_estadual_agrupado}
\title{Obter dados anuais de estoque agrupados para lista de códigos CNAE em nível estadual}
\usage{
get_estoque_emprego_estadual_agrupado(sigla_uf, nome_grupo, lista_cnae)
}
\arguments{
\item{sigla_uf}{Sigla da UF ('SP', 'RJ', etc.)}

\item{nome_grupo}{Nome para o grupo CNAE}

\item{lista_cnae}{Vetor de códigos CNAE (mesmo número de dígitos)}

\item{ano_minimo}{Ano mínimo para filtrar dados (opcional)}
}
\value{
Tibble com dados anuais agrupados por CNAE
}
\description{
Obter dados anuais de estoque agrupados para lista de códigos CNAE em nível estadual
}
