% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_estimado_estadual_anual}
\alias{get_estoque_emprego_estimado_estadual_anual}
\title{Obter dados anuais de estoque de emprego estimado em nível estadual}
\usage{
get_estoque_emprego_estimado_estadual_anual(
  sigla_uf,
  nivel_cnae = "divisao",
  codigo_cnae = NULL,
  ano_minimo = NULL,
  ano_maximo = NULL
)
}
\arguments{
\item{sigla_uf}{Sigla do estado ('SP', 'RJ', etc.)}

\item{nivel_cnae}{Nível CNAE ('divisao' ou 'grupo')}

\item{codigo_cnae}{Filtro por código CNAE específico (opcional)}

\item{ano_minimo}{Ano mínimo para filtrar dados (opcional)}

\item{ano_maximo}{Ano máximo para projeção (opcional)}
}
\value{
Data frame com coluna 'origem' indicando 'Real' ou 'Estimação'
}
\description{
Obter dados anuais de estoque de emprego estimado em nível estadual
}
