% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_estimado_estadual_anual_agrupado}
\alias{get_estoque_emprego_estimado_estadual_anual_agrupado}
\title{Obter dados anuais de estoque de emprego estimado agrupado por lista CNAE em nível estadual}
\usage{
get_estoque_emprego_estimado_estadual_anual_agrupado(
  sigla_uf,
  nome_grupo,
  lista_cnae,
  ano_minimo = NULL,
  ano_maximo = NULL
)
}
\arguments{
\item{sigla_uf}{Sigla do estado ('SP', 'RJ', etc.')}

\item{nome_grupo}{Nome para o grupo CNAE}

\item{lista_cnae}{Vetor de códigos CNAE (mesmo número de dígitos)}

\item{ano_minimo}{Ano mínimo para filtrar dados (opcional)}

\item{ano_maximo}{Ano máximo para projeção (opcional)}
}
\value{
Data frame agrupado com coluna 'origem' indicando 'Real' ou 'Estimação'
}
\description{
Obter dados anuais de estoque de emprego estimado agrupado por lista CNAE em nível estadual
}
