% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_estimado_municipal_anual_agrupado}
\alias{get_estoque_emprego_estimado_municipal_anual_agrupado}
\title{Obter dados anuais de estoque de emprego estimado agrupado em nível municipal}
\usage{
get_estoque_emprego_estimado_municipal_anual_agrupado(
  sigla_uf,
  codigo_municipio,
  nome_grupo,
  lista_cnae,
  ano_minimo = NULL,
  ano_maximo = NULL
)
}
\arguments{
\item{sigla_uf}{Sigla do estado}

\item{codigo_municipio}{Código IBGE do município}

\item{nome_grupo}{Nome para o grupo CNAE}

\item{lista_cnae}{Vetor de códigos CNAE}

\item{ano_minimo}{Ano mínimo (opcional)}

\item{ano_maximo}{Ano máximo (opcional)}
}
\value{
Data frame vazio com mensagem informativa
}
\description{
Obter dados anuais de estoque de emprego estimado agrupado em nível municipal
}
