% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_nacional_agrupado}
\alias{get_estoque_emprego_nacional_agrupado}
\title{Obter dados anuais de estoque agrupados para lista de códigos CNAE em nível nacional}
\usage{
get_estoque_emprego_nacional_agrupado(nome_grupo, lista_cnae, agregado = TRUE)
}
\arguments{
\item{nome_grupo}{Nome para o grupo CNAE}

\item{lista_cnae}{Vetor de códigos CNAE (mesmo número de dígitos)}

\item{agregado}{Se TRUE, agrega todos os estados}

\item{ano_minimo}{Ano mínimo para filtrar dados (opcional)}
}
\value{
Tibble com dados anuais agrupados por CNAE
}
\description{
Obter dados anuais de estoque agrupados para lista de códigos CNAE em nível nacional
}
