% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comex.R
\name{get_exportacao_agrupamentos_pais_ncm}
\alias{get_exportacao_agrupamentos_pais_ncm}
\title{Exportações por Agrupamento, país e NCM como tibble}
\usage{
get_exportacao_agrupamentos_pais_ncm(...)
}
\arguments{
\item{...}{Argumentos de \code{Comex$get_exportacao_agrupamentos_pais_ncm()}}
}
\value{
tibble
}
\description{
Atalho de \code{Comex$get_exportacao_agrupamentos_pais_ncm()}. Ver \code{get_exportacao_agrupamentos()}.
}
