% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comex.R
\name{get_importacao_agrupamentos}
\alias{get_importacao_agrupamentos}
\title{Importações por Agrupamento como tibble}
\usage{
get_importacao_agrupamentos(...)
}
\arguments{
\item{...}{Argumentos de \code{Comex$get_importacao_agrupamentos()}}
}
\value{
tibble
}
\description{
Atalho de \code{Comex$get_importacao_agrupamentos()}. Ver \code{get_exportacao_agrupamentos()}.
}
