% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_portal_emprego_breakdowns_estadual}
\alias{get_portal_emprego_breakdowns_estadual}
\title{Composicoes do portal (ex.: intensidade tecnologica) de uma UF}
\usage{
get_portal_emprego_breakdowns_estadual(uf, ambiente = "producao")
}
\arguments{
\item{uf}{Sigla da UF (ex.: "SP").}

\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Lista nomeada \code{{id: tibble(items)}}.
}
\description{
Composicoes do portal (ex.: intensidade tecnologica) de uma UF
}
