% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_portal_emprego_capa_estadual}
\alias{get_portal_emprego_capa_estadual}
\title{KPI de capa (\code{capa_kpi}) de uma UF}
\usage{
get_portal_emprego_capa_estadual(uf, ambiente = "producao")
}
\arguments{
\item{uf}{Sigla da UF (ex.: "SP").}

\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Lista com o KPI de capa.
}
\description{
KPI de capa (\code{capa_kpi}) de uma UF
}
