% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_portal_emprego_kpis_estadual}
\alias{get_portal_emprego_kpis_estadual}
\title{KPIs do portal de uma UF como tibble}
\usage{
get_portal_emprego_kpis_estadual(uf, ambiente = "producao")
}
\arguments{
\item{uf}{Sigla da UF (ex.: "SP").}

\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Tibble de KPIs.
}
\description{
KPIs do portal de uma UF como tibble
}
