% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_relatorio_emprego_estoque_estadual}
\alias{get_relatorio_emprego_estoque_estadual}
\title{Estoque RAIS por divisao CNAE (cache) de uma UF}
\usage{
get_relatorio_emprego_estoque_estadual(uf, ambiente = "producao")
}
\arguments{
\item{uf}{Sigla da UF (ex.: "SP").}

\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Tibble com o estoque.
}
\description{
Estoque RAIS por divisao CNAE (cache) de uma UF
}
