% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_relatorio_emprego_estoque_todos_estados}
\alias{get_relatorio_emprego_estoque_todos_estados}
\title{Estoque de todas as UFs disponiveis, empilhado, com coluna \code{uf}}
\usage{
get_relatorio_emprego_estoque_todos_estados(ambiente = "producao")
}
\arguments{
\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Tibble empilhado.
}
\description{
Estoque de todas as UFs disponiveis, empilhado, com coluna \code{uf}
}
