% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_relatorio_emprego_metadata_estadual}
\alias{get_relatorio_emprego_metadata_estadual}
\title{Metadados do relatorio (reference, fonte, gerado_em, ...) de uma UF}
\usage{
get_relatorio_emprego_metadata_estadual(uf, ambiente = "producao")
}
\arguments{
\item{uf}{Sigla da UF (ex.: "SP").}

\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Lista de metadados.
}
\description{
Metadados do relatorio (reference, fonte, gerado_em, ...) de uma UF
}
