% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_relatorio_emprego_saldo_estadual}
\alias{get_relatorio_emprego_saldo_estadual}
\title{Saldo CAGED mensal por divisao CNAE (cache) de uma UF}
\usage{
get_relatorio_emprego_saldo_estadual(uf, ambiente = "producao")
}
\arguments{
\item{uf}{Sigla da UF (ex.: "SP").}

\item{ambiente}{"producao" (default) ou "homologacao".}
}
\value{
Tibble com o saldo mensal.
}
\description{
Saldo CAGED mensal por divisao CNAE (cache) de uma UF
}
