% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_saldo_caged_municipal}
\alias{get_saldo_caged_municipal}
\title{Obter dados CAGED municipais}
\usage{
get_saldo_caged_municipal(
  nivel_cnae,
  codigos_municipio,
  codigos = NULL,
  data_minima = NULL
)
}
\arguments{
\item{nivel_cnae}{Nível CNAE ('divisao', 'grupo', 'subclasse')}

\item{codigos_municipio}{Vetor de códigos IBGE dos municípios}

\item{codigos}{Vetor de códigos CNAE opcionais}

\item{data_minima}{Data mínima em formato YYYY-MM-DD (opcional)}
}
\value{
Tibble com dados CAGED municipais
}
\description{
Obter dados CAGED municipais
}
\examples{
\dontrun{
# São Paulo capital por divisão
dados_sp <- get_saldo_caged_municipal("divisao", 3550308)

# Múltiplas cidades por grupo
dados_cidades <- get_saldo_caged_municipal("grupo", c(3550308, 3304557))
}
}
