% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_saldo_emprego_detalhado_grupos_cnae}
\alias{get_saldo_emprego_detalhado_grupos_cnae}
\title{Obter Dados de Saldo de Emprego para Grupos CNAE (Função de Conveniência)}
\usage{
get_saldo_emprego_detalhado_grupos_cnae(...)
}
\arguments{
\item{...}{Argumentos para passar para Emprego$get_saldo_emprego_detalhado_grupos_cnae()}
}
\value{
Lista com todos os dados de saldo de emprego para os grupos CNAE especificados
}
\description{
Função de conveniência para obter rapidamente TODOS os dados de saldo de emprego
para grupos nomeados de códigos CNAE sem criar explicitamente um objeto Emprego.
}
