% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_classe_cnae_estadual}
\alias{get_estoque_emprego_classe_cnae_estadual}
\title{Obter estoque de emprego por classe CNAE (4 dígitos), por UF}
\usage{
get_estoque_emprego_classe_cnae_estadual(ufs, codigos_classe = NULL)
}
\arguments{
\item{ufs}{Sigla(s) de UF (ex: 'SP' ou c('SP', 'RJ'))}

\item{codigos_classe}{Vetor de códigos de classe CNAE (opcional)}
}
\value{
Tibble com estoque por classe CNAE
}
\description{
Obter estoque de emprego por classe CNAE (4 dígitos), por UF
}
