% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_classe_cnae_nacional}
\alias{get_estoque_emprego_classe_cnae_nacional}
\title{Obter estoque de emprego por classe CNAE (4 dígitos), agregado nacionalmente}
\usage{
get_estoque_emprego_classe_cnae_nacional(codigos_classe = NULL)
}
\arguments{
\item{codigos_classe}{Vetor de códigos de classe CNAE (opcional)}
}
\value{
Tibble com estoque por classe CNAE
}
\description{
Obter estoque de emprego por classe CNAE (4 dígitos), agregado nacionalmente
}
