% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_estadual}
\alias{get_estoque_emprego_estadual}
\title{Obter dados anuais de estoque de emprego em nível estadual}
\usage{
get_estoque_emprego_estadual(sigla_uf, codigos_cnae = NULL, nivel_cnae = 2)
}
\arguments{
\item{sigla_uf}{Sigla da UF ('SP', 'RJ', etc.)}

\item{codigos_cnae}{Vetor de códigos CNAE específicos (opcional)}

\item{nivel_cnae}{Nível CNAE ('divisao', 'grupo')}

\item{ano_minimo}{Ano mínimo para filtrar dados (opcional)}
}
\value{
Tibble com dados anuais de estoque de emprego
}
\description{
Obter dados anuais de estoque de emprego em nível estadual
}
