% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_nacional}
\alias{get_estoque_emprego_nacional}
\title{Obter dados anuais de estoque de emprego em nível nacional}
\usage{
get_estoque_emprego_nacional(
  codigos_cnae = NULL,
  nivel_cnae = 2,
  agregado = FALSE
)
}
\arguments{
\item{codigos_cnae}{Vetor de códigos CNAE específicos (opcional)}

\item{nivel_cnae}{Nível CNAE ('divisao', 'grupo')}

\item{agregado}{Se TRUE, agrega todos os estados}

\item{ano_minimo}{Ano mínimo para filtrar dados (opcional)}
}
\value{
Tibble com dados anuais de estoque de emprego
}
\description{
Obter dados anuais de estoque de emprego em nível nacional
}
