% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_porte_estadual}
\alias{get_estoque_emprego_porte_estadual}
\title{Obter estoque de emprego por porte de estabelecimento e setor, por UF}
\usage{
get_estoque_emprego_porte_estadual(
  ufs,
  nivel_cnae = "divisao",
  codigos_cnae = NULL,
  porte = NULL,
  setor = NULL
)
}
\arguments{
\item{ufs}{Sigla(s) de UF (ex: 'SP' ou c('SP', 'RJ'))}

\item{nivel_cnae}{'divisao' (2 dígitos), 'grupo' (3) ou 'classe' (4)}

\item{codigos_cnae}{Vetor de códigos CNAE no nível escolhido (opcional)}

\item{porte}{Vetor de portes para filtrar (opcional)}

\item{setor}{'Indústria' ou 'Comércio e Serviços' (opcional)}
}
\value{
Tibble com estoque por porte/setor
}
\description{
Obter estoque de emprego por porte de estabelecimento e setor, por UF
}
