% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_estoque_emprego_uf_cbo}
\alias{get_estoque_emprego_uf_cbo}
\title{Obter estoque de emprego por UF, classe CNAE (4 dígitos) e ocupação (CBO)}
\usage{
get_estoque_emprego_uf_cbo(
  siglas_uf = NULL,
  codigos_classe = NULL,
  codigos_cbo = NULL
)
}
\arguments{
\item{siglas_uf}{Vetor de siglas de UF (opcional)}

\item{codigos_classe}{Vetor de códigos de classe CNAE (opcional)}

\item{codigos_cbo}{Vetor de códigos CBO (opcional)}
}
\value{
Tibble com estoque por UF/classe/CBO
}
\description{
Obter estoque de emprego por UF, classe CNAE (4 dígitos) e ocupação (CBO)
}
