% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_potec_emprego}
\alias{get_potec_emprego}
\title{Obter índice Potec (intensidade tecnológica) por classe CNAE (4 dígitos)}
\usage{
get_potec_emprego(codigos_classe = NULL)
}
\arguments{
\item{codigos_classe}{Vetor de códigos de classe CNAE (opcional)}
}
\value{
Tibble com o índice Potec
}
\description{
Obter índice Potec (intensidade tecnológica) por classe CNAE (4 dígitos)
}
