% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_renda_media_emprego}
\alias{get_renda_media_emprego}
\title{Obter remuneração média RAIS por nível de agregação e código}
\usage{
get_renda_media_emprego(tipos = NULL, codigos = NULL)
}
\arguments{
\item{tipos}{Vetor com 'Geral', 'Divisao' e/ou 'Grupo' (opcional)}

\item{codigos}{Vetor de códigos CNAE correspondentes ao tipo (opcional)}
}
\value{
Tibble com remuneração média
}
\description{
Obter remuneração média RAIS por nível de agregação e código
}
