% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_saldo_emprego_grupos_cnae_as_tibble}
\alias{get_saldo_emprego_grupos_cnae_as_tibble}
\title{Obter Dados de Saldo de Emprego para Grupos CNAE como Tibble}
\usage{
get_saldo_emprego_grupos_cnae_as_tibble(...)
}
\arguments{
\item{...}{Argumentos para passar para Emprego$get_saldo_emprego_detalhado_grupos_cnae()}
}
\value{
Um tibble com dados de saldo de emprego para os grupos CNAE especificados
}
\description{
Obter Dados de Saldo de Emprego para Grupos CNAE como Tibble
}
