% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprego.R
\name{get_saldo_emprego_lista_cnae_as_tibble}
\alias{get_saldo_emprego_lista_cnae_as_tibble}
\title{Obter Dados de Saldo de Emprego para Lista CNAE como Tibble (Função de Conveniência)}
\usage{
get_saldo_emprego_lista_cnae_as_tibble(...)
}
\arguments{
\item{...}{Argumentos para passar para Emprego$get_saldo_emprego_detalhado_lista_cnae()}
}
\value{
Um tibble com dados de saldo de emprego para os códigos CNAE especificados
}
\description{
Função de conveniência para obter rapidamente TODOS os dados de saldo de emprego
para uma lista de códigos CNAE como tibble sem criar explicitamente um objeto Emprego.
}
\examples{
\dontrun{
# Obter dados de emprego de alimentos e bebidas como tibble
tibble_comida <- get_saldo_emprego_lista_cnae_as_tibble(
  lista_cnae = c("10", "11", "12"),
  nome_grupo = "Alimentos e Bebidas",
  nivel_agregacao = "municipal",
  sigla_uf = "MG"
)
}
}
